Currently passes 72% of total
The teach gave us more tests and some time to fix them, if we do, we get +1/2 of the diference in points aquired.

The mistakes i can spot are: payments not being registered properly; client levels not working; payments and debts being switched somehow; com not being accepted for some reason.